requirements.txt
requests
python-dotenv
pandas
tqdm